Extract MethodTuple module
